tests/conftest: create module, move fixtures there
Issue #28

Please notice the useful pylint catch: "R0801: Similar lines in 2 files" (duplicate-code)
A refactoring is needed (next commit maybe)
